Fixed unit test error
